+++ py/desisim/quicksurvey.py
@@ -213,7 +213,7 @@ class SimSetup(object):
         fx.write(params.format(inputdir = self.tmp_output_path, targetdir = self.targets_path))
         fx.close()
         
-    def simulate_epoch(self, epoch, perfect=False, truth=None, targets=None, mtl=None, zcat=None):
+    def simulate_epoch(self, epoch, truth, targets, perfect=False, mtl=None, zcat=None):
         """Core routine simulating a DESI epoch, 
 
         Args:
@@ -231,12 +231,6 @@ class SimSetup(object):
             * Fiber allocation 
             * Redshift catalogue construction
         """
-        # load truth / targets / zcat
-        if truth is None:
-            truth = Table.read(os.path.join(self.targets_path,'truth.fits'))
-        if targets is None:
-            targets = Table.read(os.path.join(self.targets_path,'targets.fits'))
-            
         print("{} Starting MTL".format(asctime()))
         self.mtl_file = os.path.join(self.tmp_output_path, 'mtl.fits')    
         mtl = desitarget.mtl.make_mtl(targets, zcat)
@@ -281,7 +275,7 @@ class SimSetup(object):
         print("{} writing zcat".format(asctime()))
         newzcat.write(self.zcat_file, format='fits', overwrite=True)
         print("{} Finished zcat".format(asctime()))
-        return truth, targets, mtl, newzcat
+        return mtl, newzcat
 
 
     def simulate(self):
@@ -289,7 +283,16 @@ class SimSetup(object):
         """
         self.create_directories()
 
-        truth=targets=mtl=zcat=None
+        truth = Table.read(os.path.join(self.targets_path,'truth.fits'))
+        targets = Table.read(os.path.join(self.targets_path,'targets.fits'))
+        
+        #- Drop columns that aren't needed to save memory while manipulating
+        targets.remove_columns(['DEPTH_R', 'GALDEPTH_R'])
+        truth.remove_columns(['RA', 'DEC', 'BRICKNAME', 'SOURCETYPE'])
+        if 'MOCKID' in truth.colnames:
+            truth.remove_column('MOCKID')
+
+        mtl=zcat=None
         for epoch in self.epochs_list[self.start_epoch:]:
             print('--- Epoch {} ---'.format(epoch))
 
@@ -297,8 +300,8 @@ class SimSetup(object):
 
             self.create_fiberassign_input()
 
-            truth, targets, mtl, zcat = self.simulate_epoch(epoch, perfect=False,
-                                                            truth=truth, targets=targets, mtl=mtl, zcat=zcat)
+            mtl, zcat = self.simulate_epoch(epoch, truth, targets,
+                                            perfect=False, mtl=mtl, zcat=zcat)
 
             self.backup_epoch_data(epoch_id=epoch)